+++ sys/netgraph/ng_deflate.c
@@ -39,8 +39,7 @@
 #include <sys/endian.h>
 #include <sys/errno.h>
 #include <sys/syslog.h>
-
-#include <net/zlib.h>
+#include <sys/zlib.h>
 
 #include <netgraph/ng_message.h>
 #include <netgraph/netgraph.h>
