+++ sys/sys/zutil.h
@@ -17,7 +17,7 @@
 #define ZEXPORT
 
 #ifdef _KERNEL
-#include <net/zlib.h>
+#include <sys/zlib.h>
 #else
 #include "zlib.h"
 #endif
